Refactor: Moved Step generation to a service file
